[RO-4342] Update ELK and dependencies for Pike
ELK and related dependencies need to be updated for the Pike release.
The initial commit updates the Elasticsearch packages to the latest
6.1.3 release.

Issue: RO-4342

Generate RPCO Secrets

The RPCO secrets example file is not being copied over as the default
yaml file, nor are the secrets being generated.

Typo fix