ASC-1371 Assign SYS_TEST_BRANCH via RE vars
This commit expands the conditional for assigning the value of
`SYS_TEST_BRANCH`.  PM jobs use `RE_JOB_BRANCH`, while PR jobs use
`RPC_GATING_BRANCH` to provide the branch value to the job.  These
parameters are inspected in turn to assign the value of
`SYS_TEST_BRANCH` to ensure that the system-tests branch is matched to
the RPC-Openstack version under test.