Update flavor and image setup
This change adds newer images and sets the playbook up to only
deploy images for a given virt-type, which will speed up the
general gate times and test cycles as only the images using the
supported virt type will be enabled.

The deploy scripts have been updated to default to qemu (the
current default) when deploying an AIO. When the playbooks are
executed outside of the gate environment the default will be
KVM, which is the openstack default.

Signed-off-by: Kevin Carter <kevin@cloudnull.com>