[RI-573] Update ELK MNAIO deployment
This updates the ELK MNAIO deployment methodology to use the new
MTC tooling developed by the Tools team.

This also updates the OSA test repository to a known good SHA.  The
SHA will need to be updated once master moves to T.

JIRA: RI-573

Update to use git

This uses git and the MTC_BRANCH variable to allow the user to
specify a SHA for the MTC deployment.