Enable snapshot image creation for PUSH and TIMER
To ensure that the snapshot image creation happens more regularly,
we ensure that the preparation for it happens for both a PUSH and
a TIMER trigger.

This will ensure that the images are updated on a daily basis,
not just when a new patch merges.

JIRA: RE-2278
(cherry picked from commit e11f62aa8b2ee940619306378ddec935e58d9af8)