Points Newton to use openstack_hosts stable/newton
Changes newton branch to utilize openstack_hosts
stable/newton branch instead of a sha to pull in
https://github.com/rcbops/openstack-ansible-openstack_hosts/pull/4
and so we can avoid having to modify rpc-openstack in case
we need to make future changes to openstack_hosts for newton.

The change mainly drops off the linux-extra-* from being installed
due to the issues we were having with kernel version mismatches.